move package AccessPlayerDTO
