+++ src/main/java/com/wayvi/wfly/wflyv2/api/FlyManager.java
@@ -1,9 +1,7 @@
 package com.wayvi.wfly.wflyv2.api;
 
-import com.wayvi.wfly.wflyv2.storage.AccessPlayerDTO;
 import org.bukkit.entity.Player;
 
-import java.sql.SQLException;
 import java.util.UUID;
 
 public interface FlyManager {
